# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QueryGenie API
  description: |
    Enterprise-grade API for QueryGenie - AI-powered database assistant with multi-role collaboration.
    
    ## Authentication
    All API endpoints require authentication via JWT tokens obtained through NextAuth.js.
    
    ## Rate Limiting
    - **Free Tier**: 100 requests per 15 minutes
    - **Pro Tier**: 1000 requests per 15 minutes  
    - **Enterprise**: Custom limits
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format.
    
  version: 1.0.0
  contact:
    name: QueryGenie API Support
    url: https://querygenie.com/support
    email: support@querygenie.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.querygenie.com
    description: Production server
  - url: https://staging-api.querygenie.com
    description: Staging server
security:
  - BearerAuth: []
  - SessionAuth: []

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /workspaces/{workspaceId}/members:
    get:
      tags: [Workspaces]
      summary: List workspace members
      description: Returns all members of a workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: List of workspace members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkspaceMember'
                  total:
                    type: integer
    
    post:
      tags: [Workspaces]
      summary: Invite user to workspace
      description: Sends an invitation to join the workspace (requires admin+ role)
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserRequest'
      responses:
        '201':
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /workspaces/{workspaceId}/connections:
    get:
      tags: [Connections]
      summary: List database connections
      description: Returns all database connections in a workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: List of database connections
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Connection'
                  total:
                    type: integer
    
    post:
      tags: [Connections]
      summary: Create database connection
      description: Creates a new database connection (requires admin+ role)
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectionRequest'
      responses:
        '201':
          description: Created database connection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token

  # Common Parameters
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        avatar:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [admin, developer, analyst]
        emailVerified:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    WorkspaceRole:
      type: string
      enum: [owner, admin, editor, viewer]
      description: |
        - **owner**: Full access including deletion
        - **admin**: Manage connections, members, settings
        - **editor**: Create/execute queries, edit docs  
        - **viewer**: Read-only access
...
    WorkspaceMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        isActive:
          type: boolean
        joinedAt:
          type: string
          format: date-time
    InviteUserRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/WorkspaceRole'
    Invitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        status:
          type: string
          enum: [pending, accepted, declined, expired]
        invitedBy:
          $ref: '#/components/schemas/User'
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # Connection Types
    ConnectionType:
      type: string
      enum: [postgres, mysql, mssql, sqlite]
    ConnectionMode:
      type: string
      enum: [read-only, read-write]
    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/ConnectionType'
        mode:
          $ref: '#/components/schemas/ConnectionMode'
        host:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        database:
          type: string
          nullable: true
        isActive:
          type: boolean
        lastConnectedAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
...
    CreateConnectionRequest:
      type: object
      required:
        - name
        - type
        - credentials
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          $ref: '#/components/schemas/ConnectionType'
        mode:
          $ref: '#/components/schemas/ConnectionMode'
        host:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        database:
          type: string
        username:
          type: string
        password:
          type: string
          description: Will be encrypted
        sslConfig:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
        title:
          type: string
          description: A short, human-readable summary of the problem type
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          description: Request trace ID for debugging
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
                  value:
                    description: The invalid value

  # Common Responses
  responses:
...
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: https://querygenie.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action
            instance: /api/workspaces/123/connections
            timestamp: "2024-01-20T10:30:00Z"
            traceId: abc123
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            type: https://querygenie.com/errors/validation
            title: Validation Failed
            status: 400
            detail: One or more fields contain invalid values
            instance: /api/workspaces
            timestamp: "2024-01-20T10:30:00Z"
            traceId: abc123
            errors:
              - field: name
                code: required
                message: Name is required
              - field: slug
                code: pattern
                message: Slug must contain only lowercase letters, numbers, and hyphens
                value: "Invalid Slug!"
...